refactoring swagger config
