+++ conf/swaggerSetup.js
@@ -1,0 +1,54 @@
+const swaggerUi = require('swagger-ui-express')
+const swaggerJSDoc = require('swagger-jsdoc')
+const package = require('./../package.json')
+require('dotenv').config()
+let host
+if (process.env.NODE_ENV == 'testnet') {
+    host = process.env.BASEURL
+} else if (process.env.NODE_ENV == 'local') {
+    host = process.env.BASEURLLOCAL
+} else {
+    host = process.env.BASEURL_MAINNET
+}
+const swaggerDefinition = {
+    openapi: '3.0.0',
+    info: {
+        title: 'API for node-satt',
+        version: package.version,
+        description:
+            'Welcome to SaTT Webservice endpoint, this backend provides webservice to SaTT WebWallet and advertising campaign manager',
+        customCss: '.swagger-ui .topbar { display: none }',
+    },
+    host: host,
+    components: {
+        securitySchemes: {
+            bearerAuth: {
+                type: 'http',
+                scheme: 'bearer',
+                bearerFormat: 'JWT',
+            },
+        },
+    },
+    security: [
+        {
+            bearerAuth: [],
+        },
+    ],
+}
+
+const cssOptions = {
+    customCss: `
+    .topbar-wrapper img {content:url(/assets/SaTT.png); width:50px; height:auto;}`,
+    customSiteTitle: 'SaTT',
+    customfavIcon: '/assets/SaTT-noire.png',
+}
+
+const options = {
+    swaggerDefinition,
+    apis: ['./routes/*.js'],
+}
+const swaggerSpec = swaggerJSDoc(options)
+
+
+
+module.exports = {swaggerUi, swaggerSpec, cssOptions};
